pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prasword"
version = "0.1.0"
description = "Professional word processor for academic and data science workflows"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "PrasWord Contributors" }]

dependencies = [
    "PySide6>=6.6",
    "python-docx>=1.1",
    "pygments>=2.17",
]

[project.optional-dependencies]
academic = [
    "bibtexparser>=1.4",
]
datascience = [
    "sympy>=1.12",
    "pandas>=2.0",
]
git = [
    "gitpython>=3.1",
]
all = [
    "bibtexparser>=1.4",
    "sympy>=1.12",
    "pandas>=2.0",
    "gitpython>=3.1",
]
dev = [
    "pytest>=7.4",
    "ruff>=0.1",
    "mypy>=1.6",
]

[project.scripts]
prasword = "prasword.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["prasword*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
line-length = 99

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
